feat(ui): 트리/그리드 이미지 버전 우측 병치 + Border 0.5px + 버튼 아이콘 16
- 갤러리: TreeView·TreeGrid의 '행 이미지' 버전을 각각 우측에 나란히 배치(좌 기본/우 이미지).
  라벨을 컨트롤 x에 맞춰 그려 좌·우 쌍 라벨 분리. content_w는 가장 넓은 행 기준(가로 스크롤).
- BorderSpec.width i32→f32 — 0.5px 소수 두께 허용(SDF AA 얇은 선). 트리 테두리 0.5px.
- 버튼 선행 이미지 16(콤보 18 − 사방 1px).

ui 96 green. clippy·fmt clean.

+++ crates/nbeep-ui/src/controls/mod.rs
@@ -33,11 +33,11 @@ use crate::draw::{DrawCtx, FontSlot};
 use crate::geom::{Point, Rect};
 use crate::theme::{Color, Theme};
 
-/// 외곽 테두리 설정 — 두께(논리 px)·색·투명도(0..=1). 두께 0 = 테두리 없음(기본).
+/// 외곽 테두리 설정 — 두께(논리 px · 소수 가능 예: 0.5)·색·투명도(0..=1). 두께 ≤0 = 없음(기본).
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct BorderSpec {
-    /// 두께(논리 px). 0이면 그리지 않는다.
-    pub width: i32,
+    /// 두께(논리 px). 0 이하면 그리지 않는다. 0.5 같은 소수도 허용(SDF AA 얇은 선).
+    pub width: f32,
     /// 색.
     pub color: Color,
     /// 불투명도(0.0~1.0).
@@ -47,7 +47,7 @@ pub struct BorderSpec {
 impl Default for BorderSpec {
     fn default() -> Self {
         Self {
-            width: 0,
+            width: 0.0,
             color: Color::from_rgb(0x36, 0x3C, 0x46),
             alpha: 1.0,
         }
@@ -57,7 +57,7 @@ impl Default for BorderSpec {
 impl BorderSpec {
     /// (두께, 색, 투명도)로 만든다.
     #[must_use]
-    pub fn new(width: i32, color: Color, alpha: f32) -> Self {
+    pub fn new(width: f32, color: Color, alpha: f32) -> Self {
         Self {
             width,
             color,
